Refactor API layer: add Vercel serverless support, CORS, dark mode fix
- Extract shared logic to api/_shared.js, remove duplication from server.js
- Add Vercel serverless functions (api/analyze.js, api/providers.js)
- Add CORS helper with ALLOWED_ORIGIN env var support
- Sanitize 500 error messages to prevent internal detail leakage
- Add vercel.json with API-aware SPA rewrite rules
- Fix dark mode form input visibility (color-scheme, inherit background)
