+++ api/providers.js
@@ -1,0 +1,12 @@
+import { getAvailableProviders, setCorsHeaders } from './_shared.js';
+
+export default function handler(req, res) {
+  setCorsHeaders(res);
+  if (req.method === 'OPTIONS') return res.status(200).end();
+  if (req.method !== 'GET') {
+    res.status(405).json({ error: 'Method not allowed' });
+    return;
+  }
+
+  res.status(200).json(getAvailableProviders());
+}
